Move BroadcastReceiver from ViewModel to View
