+++ V2rayNG/app/src/main/kotlin/com/v2ray/ang/ui/MainActivity.kt
@@ -58,7 +58,43 @@ class MainActivity : BaseActivity(), NavigationView.OnNavigationItemSelectedList
     }
     private var mItemTouchHelper: ItemTouchHelper? = null
     val mainViewModel: MainViewModel by viewModels()
+    private val mMsgReceiver = object : BroadcastReceiver() {
+        override fun onReceive(ctx: Context?, intent: Intent?) {
+            when (intent?.getIntExtra("key", 0)) {
+                AppConfig.MSG_STATE_RUNNING -> {
+                    mainViewModel.isRunning.value = true
+                }
 
+                AppConfig.MSG_STATE_NOT_RUNNING -> {
+                    mainViewModel.isRunning.value = false
+                }
+
+                AppConfig.MSG_STATE_START_SUCCESS -> {
+                    toast(R.string.toast_services_success)
+                    mainViewModel.isRunning.value = true
+                }
+
+                AppConfig.MSG_STATE_START_FAILURE -> {
+                    toast(R.string.toast_services_failure)
+                    mainViewModel.isRunning.value = false
+                }
+
+                AppConfig.MSG_STATE_STOP_SUCCESS -> {
+                    mainViewModel.isRunning.value = false
+                }
+
+                AppConfig.MSG_MEASURE_DELAY_SUCCESS -> {
+                    mainViewModel.updateTestResultAction.value = intent.getStringExtra("content")
+                }
+
+                AppConfig.MSG_MEASURE_CONFIG_SUCCESS -> {
+                    val resultPair = intent.getSerializableExtra("content") as Pair<String, Long>
+                    MmkvManager.encodeServerTestDelayMillis(resultPair.first, resultPair.second)
+                    mainViewModel.updateListAction.value = mainViewModel.getPosition(resultPair.first)
+                }
+            }
+        }
+    }
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         binding = ActivityMainBinding.inflate(layoutInflater)
@@ -159,9 +195,24 @@ class MainActivity : BaseActivity(), NavigationView.OnNavigationItemSelectedList
             }
             hideCircle()
         }
-        mainViewModel.startListenBroadcast()
+        mainViewModel.isRunning.value = false
+        startListenBroadcast()
+        MessageUtil.sendMsg2Service(application, AppConfig.MSG_REGISTER_CLIENT, "")
     }
-
+    fun startListenBroadcast() {
+        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.TIRAMISU) {
+           registerReceiver(
+                mMsgReceiver,
+                IntentFilter(AppConfig.BROADCAST_ACTION_ACTIVITY),
+                Context.RECEIVER_EXPORTED
+            )
+        } else {
+           registerReceiver(
+                mMsgReceiver,
+                IntentFilter(AppConfig.BROADCAST_ACTION_ACTIVITY)
+            )
+        }
+    }
     private fun copyAssets() {
         val extFolder = Utils.userAssetPath(this)
         lifecycleScope.launch(Dispatchers.IO) {
@@ -697,5 +748,10 @@ class MainActivity : BaseActivity(), NavigationView.OnNavigationItemSelectedList
         }
         binding.drawerLayout.closeDrawer(GravityCompat.START)
         return true
+    }
+
+    override fun onDestroy() {
+        super.onDestroy()
+        unregisterReceiver(mMsgReceiver)
     }
 }
